New functionality COLLECT MESSAGES added
This funcionality allows to defer errors until the collection of messages is stopped. It works similar to bulk exceptions in PL/SQL but is aimed at validation methods which perform mulitple validations. By collecting those messages, it's possible to retrieve multiple errors instead of calling the validation code multiple times