Method LOG_STATE added
This method allows to log an arbitrary list of parameter or variable values at any place in the code. It takes an instance of MSG_PARAMS as in the ENTER or LEAVE method but doesn't change the call stack hierarchy.

It also uses an additional parameter called LOG_STATE_THRESHOLD that sets the severity from when on PIT wil emit LOG_STATE messages.