MSG_PARAM enhanced by constructor function
When entering a method and providing parameter values to the enter/leave or log_state methods, a limited of varchar2(4000) applies for the parameter value. To circumvent that this type throws exception if this length is exceeded, a constructor method was provided that limits the length accordingly, regardless of the length of the parameter value passed in.