New install scripts, dependency from global context made optional
The new install scripts allow for DBA-less installation based on BAT or SH scripts.
Plus, it is no longer required to be able to habe a global context in place. In the vicinity of APEX, there is no need for a global context so at least here it is now possible to use all features of PIT even if no global context is present.